+++ R/utils.R
@@ -15,7 +15,7 @@
 speMCA_varsup <- function(mca, df) {
     if (!is.data.frame(df)) stop("df must be a data frame")
     res <- lapply(names(df), function(name) {
-        l <- GDAtools::varsup(mca, df[, name])
+        l <- GDAtools::supvar(mca, df[, name])
         l <- lapply(l, function(x) {
             if (is.data.frame(x)) {
                 rownames(x) <- paste(name, rownames(x), sep = "____")