+++ NEWS.md
@@ -1,5 +1,6 @@
 # explor (development version)
 
+- Fix varsup and indsup functions renamed to supvar and supind in GDAtools
 - Fix duplicated level names in supplementary variables in speMCA
 - Fix warnings in Font Awesome icon names (#39, thanks @jl5000)
 - Fix supplementary variables not displayed in individual plot for speMCA (thanks @419kfj)